Answer the third review round on member keys and suite shape
A reserved word keeps its own shape only in keyword position: after a
property-access dot or in an object-key slot it is a member name, so
{ delete: first } and { archive: second } read as one renamed pair while
case and default stay keywords. Normalize carriage returns before
hashing, so a CRLF checkout reads the LF registry. Validate every
registry entry against a schema, so a hand-edit without a reason cannot
ride its hash through the check. Decode the repository root through
fromFileUrl. Keep the slug-writes contract and the provider slug builder
private, and pin their public behaviour with exact-value tests. Refresh
the clone counts, and split the api-groups suite into focused files with
the duplicate assertion removed.
